Implemented authentication on ims_data_client

+++ queng_message/message_stream/src/client_stream.rs
@@ -19,7 +19,7 @@ impl MessageStream {
     }
     async fn build(client_id: u16) -> Result<Self, Error> {
         let user = IggyUser::default();
-        let iggy_config = IggyConfig::from_client_id(user, client_id as u32);
+        let iggy_config = IggyConfig::from_client_id(user, client_id);
         let stream_id = iggy_config.stream_id();
         let topic_id = iggy_config.topic_id();
